+++ bin/rails
@@ -1,0 +1,8 @@
+#!/usr/bin/env ruby
+# This command will automatically be run when you run "rails" with Rails 4 gems installed from the root of your application.
+
+ENGINE_ROOT = File.expand_path('../..', __FILE__)
+ENGINE_PATH = File.expand_path('../../lib/tagger/engine', __FILE__)
+
+require 'rails/all'
+require 'rails/engine/commands'
